Kill threads during shutdown.
Take advantage of the fact that CM socket accept and receive
calls return EINTR if the thread they are in is killed.
Use pthread_kill() and check the return value of such APIs.
If EINTR then terminate the thread gracefully.
Wait in main thread for all threads to terminate before
completing the shutdown
Handle SIGUSR1 events (by doing nothing) as they are the ones
used in pthread_kill() and must be handled