Compiler error fixups and other cleanup
This commit allows cross-compilation of ARM to succeed, without
warnings. Raw compiler invocations have been replaced with the proper
variable "CC" or "CXX", which are themselves defined in the top-level
makefile.

Gitignore files have been expanded or added to ignore the binary files
that are created during compilation.

A linker warning caused by the variable ptl_all_ports being declared
twice has been fixed (by removing the one in the header file), this
variable isn't actually used anywhere else so this is a safe change for
the time being.

When printing 64 bit values, the format string will be different on 32
and 64 bit machines, therefore we need to use the PRIu64 macro provided
by inttypes.h to ensure portability.

The rdma/Makefile has also had the flag -pthread passed to its linker
command line, as the pthread library is not automatically linked in some
platforms and must be done explicitly.

Finally a spurious compiler error caused by the char being signed on
some platforms and unsigned on others was fixed, by using an int for the
return of getopt in two file.