Switch to using common static libraries and static RDMA library
Still have issues with building rdma/samples (but not rdma/test).
I have temporarily taken the rdma/samples out of the build.
- Still using libmport.so for now. Should we create a separate directory
for libmport.a? When both .so and .a are in the same directory the .so
takes precedence.
- No longer building librdma.so by taking it from the 'all' target, but
leaving the build command in the Makefile in case we need it.