Default log level set to 7. Other defaults eliminated.
The log level is set to 7 (DBG) in the master Makefile and exported
as an environment variable. Other (child) makefiles may use that value
or override it on the command line if building their component separately.
The DEBUG flag is only relevant to liblog since it just determines whether
to echo the log messages to the screen.