For now fail if we try to provision an already known destid
This is just to keep things simple for now. Once they work we should probably
update the destid info (i.e. replace the entry in the list, kill the old
thread and create a new one...etc. Just that the implications of doing so while
there maybe active connections would need more thought.