Connection/disconnection should use socket if created by the provisioning daemon as well.
The socket representing connection to a particular destid's daemon might have
been created in the prov thread (in response to a HELLO message), or
in the provision command (which sends a HELLO message). The former is a server 'accept'
socket represented by a cm_server object, while the latter is a client socket, represented
by a cm_client object. Hence we use polymorphism and store a pointer to cm_base so
the connection/disconnection code would work for either types.
TO BE TESTED.