Do not close previous accept sockets in the accept() method. Thanks Barry.
Each accept_socket represents a connection to a remote daemon and is passed to
a 'child' cm_server object that will maintain it and do send/recv on it.
The closure of the socket was done to prevent leaking sockets if a second
riodp_socket_accept() was called causing the prior socket to be overwritten and
hence not properly closed. This is no longer the case since 'child' cm_server
objects will close their respective accept sockets upon destruction. The parent
just has to take care of the one and only accept socket which has not accepted
 a connection yet, and that one, experience shows, actually doesn't need to be
closed.