FIX test case. The RDMA connect was getting ahead of the accept.
The test case tries to put the server in RDMA-accept mode for a particular memory
space. Then it sends an RDMA-connect request. The problem is that the 'connect'
gets ahead of the 'accept' and the server discards the connection request to that memory
space since the memory space isn't listed among the memory spaces awaiting a connection.
Solution: Insert a delay that better guarantees that the BAT server has executed the RDMA
accept before the RDMA-connect message arrives.