Switch to always use msg_q wrapper. Proper cleanup in wait_for_disc_thread
Always use the class wrapper msg_q since it does proper cleanup based on whether
the object owns the queue or not.
wait_for_disc_thread_f() now properly marks the message queue and the thread
as unused in the database entry. Furthermore, locating the memory space
is based on the message queue name and not on the memory subspace identifier.