Distinguish between UNIX server sockets created by the 2 different methods
In one method using the fully-fledged constructor we also perform an accept()
to obtain an accept socket. Therefore upon deletion we must also close 'the_socket'.
With the minimal constructor the object is constructed and passed an accept_socket
which can be used directly for send/recv.
On destruction the behaviour is different as in the first case we need to close
'the_socket' while in the latter we SHOULD NOT do so as it can be used to accept
further connections.
With this change we can delete the server object for any RDMA application that
quits or crashes thus preventing a memory leak.