Retry communication with dead daemon on each RDMA API call.
Upon failing to communicate with a daemon the library closes the socket and
indicates that it is not fully initialized. Upon calling rdma_lib_init()
or any other RDMA API, a socket connection to the daemon is retried.
This way an application does not have to explicitly re-init the RDMA lib.