Make sure RDMA daemon is alive before mapping/unmapping
APIs such as rdma_unmap_msub() and rdma_map_msub() don't call the daemon
but if the daemon is dead the physical inbound space isn't allocated
and therefore would not be mappable/unmappable.
Added a dummy call rdmad_is_alive() between the library and the daemon
that does nothing but verify that the daemon is alive.