Recovery from restarting the RDMA daemon.
Any API that needs the daemon first checks for its presence. In doing so, it
will also purge the local database if the daemon had died, and sets the init flag
to false. The API then calls CHECK_LIB_INIT() which attempts to re-connect to the daemon.
If that works the rest of the code in the API proceeds.