Distinguish between log level and display level.
With 'levelog' we limit what goes into the debug buffer
With 'leveldisp' we limit what gets displayed on the screen.
The goal is to set levelog to 7 and leveldisp say to 3 or 4.
This way we watch the screen and when a warning or error shows up we
do 'dlog' and see the context around the warning from the buffer.