BUGFIX: Don't 'abort' if an 'ms' was destroyed before a 'disconnect' message arrives.
This happens in current (old) test case 'i'. The client exits and the daemon decides
to clean up after the client by trying to disconnect from the client's remote 'ms'.
Meanwhile, the remote 'ms' was destroyed before the 'disconnect' message arrives.
This is not an error; the 'disconnect' message is just ignored.