Properly propagate result of destroying a memory space/subspace to caller
As it turns out when a functor is passed to for_each() a copy is instantiated
and therefore one cannot use a public member of the object to pass results
back to a caller. One could pass a 'result' variable by reference to the constructor
of the functor. However, since this code is only used once, an inline Lampda function
is the way to go.