"owners" no longer a global referenced from various class methods!
It is poor design to have a method in the memory space class, for example,
refer to a global variable. Now when the inbound space is created, it is
passed a reference to the 'owners' object. The inbound class makes this
reference available to the inbound windows. The 'owner's object is now
local to rdmad_main.cpp