BUGFIX: No need to kill thread when removing daemon entry
A daemon entry is removed when the remote daemon dies. Since
the local daemon establishes two CM socket connections with
the remote daemons, the remote daemon's death is detected and
the threads holding the connections self-exit after destroying
the socket. Therefore none of this needs to be done in remove_daemon()