Remove redundant daemon validation code
A DID (destid) isn't added to any of the 2 lists (prov and hello) unless
there is a daemon running on that node. Then when the remote daemon is
detected, 2 connections are established with that daemon. Should the
daemon die, the connections are broken, the sockets are destroyed,
the daemons are removed from the relevant lists (prov / hello) from
within the threads holding the connection. We don't need explicit
code to check if a remote daemon is still running.: