Rethought handling dropped daemons/nodes and shutting down
- No need to explicitly check for dropped daemons as we know when
the CM connection is "reset by peer".
- When 'reset by peer' then we must delete the socket, remove the
daemon entry, and exit the thread. The thread is now dead.
- On shutdown: we explicitly kill the thread. We know from the return
code of the CM socket receive (EINTR) that the thread was killed due
to shutdown. We delete the CM socket but don't bother with removing
the daemon entry; the daemon list is global and will go out of scope
anyway. The important thing is that the thread was killed and the
socket was closed.