libriocp_pe: Catch uninitialized switch driver when handling an event
The handle event function needs to check for the switch driver to be
initialized when trying to handle an event.
Otherwise this would cause an segmentation fault. When the uninitialized
driver is catched a log is given. The riocp_pe_switch_handle_event()
waits 20x100ms unfitl it returns -ETIMEDOUT.