BUGFIX: Destroy mspaces and coalesce free space when connection drops
This was discovered when RSKTD was closed. The memory spaces were freed
but the freed space was not coalesced (merged). Turns out the correct function
was not called.
Also in getspaces_connected_by_desid() a 'this->' qualifier was missing
which is needed to distinguish the member 'mspaces' from the parameter
'mspaces'