tx_engine::send_message() now makes a copy of the message before queuing
Noticed when testing with RSKT that some messages were corrupted between
the initial reception and actually handling their contents. This is probably
because the messages were going out of scope. Now we make a copy and queue
a copy of the message in the message queue, and we don't have to make
the original message buffer static anymore.