libriocp_pe: enable event log for IDT gen2 switches
The event log of the IDT gen2 switches stores up to 256 events. It is
cleared by readout. The event log implementation will read out the log
every time when a port write event has been received and the result is
printed to the syslog. With that the event log outputs should be related
to the events.

It could happen that event log items are printed for a following event
when this has been triggered in the switch while a previous event is
handled by the host.