Don't add remote connection to the memory space until server responds
If the server app fails to receive the CONNECT_MS_REQ it will try
to UNDO_ACCEPT for the specified memory space. However, it the memory
space has a remote connection for that server app, it cannot UNDO_ACCEPT.
The solution is to defer adding the remote connection and switching the
memory space state from is_accepting to connected_to until the server
sends CONNECT_MS_RESP to the server daemon.