Fix crash on quit
There were several problems:
- Had to use pthreads with the Tx and Rx engines (especially Rx engine). C++
threads (std::thread) don't accept termination via pthread_kill(). While they
do provide thread::native_handle which in theory should be pthread_t in reality
it is always 0.
- When quitting, we must clear both daemon lists unconditionally (not just engines
marked as dead). This was nasty as it left some engines alive when we shutdown
causing the destructor to access sockets that have died, trying to print to
a log file that was closed...etc.

This changelist contains unnecessary changes probably. I've changed many logging
to printf to get to the bottom of the problem. I will revert those changes
coming forward. The important thing is we don't crash on 'quit' anymore and we seem
to be cleaning up everything in an orderly manner as far as I can tell.