libriocp_pe: cps1xxx_get_route_entry: fix 8-bit destid handling
cps1xxx_get_route_entry() did not work with 8-bit destination
ids because it always tried to do the lookup in the domain routing
table first, and for domain 0 it obtained 0xde.

Since the IDT gen2 switches have an implicit handling for domain
0 (destination is handled as an 8-bit destination id), change
this function to perform the lookup only in the device routing
table for such 8-bit destination ids.