Increase backtrace size and also exit on segfaults.
While it is nice to shut everything down, experience shows that when
the daemon is in a bad state anyway trying to shutdown (and in the process
trying to close all remote connections, notify remote servers, clean up
all memory spaces...etc.) is also likely to fail and procude more segfaults
that loop over and over again. Better to quit after printing out the backtrace
so it is clear what caused the segfault.