BUGFIX: Distinguish between server-app-initiated and forced disconnections
A server-app-initiated disconnection of a memory space is made by calling
rdma_disc_ms_h() with server options. The API blocks waiting for confirmation
that remote clients have received the request and acted on it.
There needs to be at least 1 remote client currently connected to the
memory space or the API will eventually timeout waiting for the acknowledgement.
As such the request goes from server-app to server-daemon to client-daemon to
client-app, then back. The message needs to contain the server-side tx_eng
so that the server daemon can deliver the ack.

Forced disconnections are different. They mainly occur if:
1. The server-app destroys the memory space, or
2. The server-app dies
In both cases the remote daemon is sent a CM_FORCE_DISCONNECT_MS and
no acknowledgement is expected since either the memory space is already
dead or the server app is dead and cannot process any acknowledgement

The bug was that in trying to forward the server disconnection acknowledgement
to the server app, the tx_eng used was that of the client app thus producing
a segmentation fault and killing the daemon. This caused test case 'j' to fail.