Replace await_message() with explicit notification configuration
await_message() is flawed since it assumes the action prior to calling it
would be slow enough as to allow await_message() to call set_notify()
before any response to that action arrives. During testing it was shown
that this sometimes happens. Now we always do set_notify(), perform the
action, then wait for the notification semaphore to be set.