BUGFIX: Support multiple event notifications from different threads
When a message arrives and if it is of type RDMA_CALL, we compare
the type, category, and sequence number with the notify list paramters.
This is because an incoming message will not include a semaphore.
However, when setting a notification, we should NOT consider a notification
parameter to be duplicate UNLESS it is has the same type, category,
sequence number, AND SEMAPHORE. This was breaking test case 'm'
as trying to set multiple notifications from different threads
was only setting a single notification and discarding the others
as duplicates.