libriocp_pe: add global synchronization primitives
This patch adds a global mutex and API functions to lock and unlock it.
The mutex is mainly meant for synchronizing riocp_pe_event_mport() with
everything else. Because waiting for the next available event is done
from inside riocp_pe_event_mport(), the mutex cannot be implemented
outside libriocp_pe.

Signed-off-by: Radu Rendec <radu.rendec.ext@nokia.com>