libriocp_pe: add get supported port speeds functionality
The LP-Serial Lane Extended Features Block is a standard block, but
unfortuntely some of the registers, such as "Lane n Status 3 CSR" are
implementation specific. Because of that, the implementation of this
functionality has to go through a switch driver specific callback.

The generic part only checks if this extended feature block (id = 0xD) is
supported or not by the target device.