libriocp_pe: Improve los of lane trigger at port disable
Problem: For a link down notification the los of lane trigger must trigger.
For that during the link down handling the speed rate configuration is "misconfigured"
to trigger that event. This seems not always working and only descrambler
and idle2 out of sync is detected by the SERDES logic, but no los.

Fix: now manually trigger the los of lane condition by writing the
los bits in the lane error detect register. This will fire the los of lane trigger.
A small wait time is needed to give the port logic to send a port write before it
is disabled. That can be done by waiting por the port write pending status.