Add "port self multicast" control function
The riocp_sw_set_port_self_mcast() function provides control over the
"self multicast" feature at the port level. This is an IDT specific
feature and returns -ENOSYS for other switch types.

Feature description from the IDT switch manual:
If this port is included in any multicast mask, multicast
packets received on this port will be multicast back out this port. If
this bit is set, software must handle multicast back to itself. Note
that multicasting packets back to the port that the packet was
received on is contrary to the RapidIO Specification (Rev. 2.1).