libriocp_pe: preserve PORT_X_CTL_1_CSR in cps1xxx_lock_port and cps1xxx_unlock_port
cps1xxx_lock_port and cps1xxx_unlock_port were overwriting other bits in the
PORT_X_CTL_1_CSR register. This caused problems in some situations. For
example when disabling a port, we set the CPS1xxx_CTL_PORT_DIS bit in this
port but we also trigger a los of lane. When the event is received, the event
handler would clear CPS1xxx_CTL_PORT_DIS when calling cps1xxx_lock_port.

This patch uses a read/modify/write approach to conserve the other bits in
this register.