libriocp_pe: Fix wrong route table selection bug
The driver implementation inplicitly expected that switch destid(lsb) is
always 0x00. When that is not the case a domain route is programmed
instead of a device route for destid 0x<dm>00. When the route
destination domain ID is the same as the switch domain ID the switch
implicitly programs then 256 device routes for that (already existing
domain).