fixed port speed and enable/disable handling
Fixed the idtgen2 functions to follow the "no port reset while link up"
rule. This avoids a race condition where the speed setup establishes a
link that needs to be reported by port write. A port reset after the
link up might cause the port write to be cancelled, by that the link up
is not detected. The port reset is moved to the link down handler, there
it is safe to call it.

Signed-off-by: Frank Kunz <frank.kunz@nokia.com>