libriocp_pe: move comptag allocation outside of the library
This patch completely removes comptag allocation code from the library
and leaves comptag allocation to the user code.

Comptag separation into host id (lower 16 bits) and number (next 8 bits)
is also removed from the library and the comptag is treated as an opaque
32-bit value.

However, a comptag mask is still supported, but provided from outside
the library when the mport handle is created. The mask is used by the
library for:
- configuring the port write filters when a new port write listener is
  registered with the mport driver;
- clearing the unused bits when the comptag value is extracted from the
  raw port write data;
- clearing the unused bits when the comptag is read from the endpoint
  comptag register;
- preserving the unused bits when the comptag is updated in the endpoint
  comptag register.

The mask allows the user code to reserve some of the comptag bits for
other purposes than the actual comptag data.