Removal of duplicate/deprecated constant definitions and macros from RapidIO_Routing_Table_API.h, in favour of rio_standard.h and rio_ecosystem.h values.  Constants removed are:
Macros removed are:

The RIO_DSF_MAX_MC_MASK change has the impact of increasing the maximum
number of multicast masks from 255 to 256.  This has the follow on impact
of increasing some multicast mask index types from uint8_t to uint32_t,
and altering some tests to avoid the serendipitous side effect of using
a contastant of 0x100 instead of 0xFF.

Also found/fixed a bug with RXS RT support:  RXS does not support the
Multicast Mask info register, so generic support for multicast mask allocation
had to be split to support a "maximum masks" parameter.